rename slider
