+++ src/components/CitationExporter/components/index.tsx
@@ -1,6 +1,6 @@
-export * from './AuthorCutoffSlider';
+export * from './AuthorCutoffField';
 export * from './DescriptionCollapse';
 export * from './FormatSelect';
 export * from './JournalFormatSelect';
 export * from './KeyFormatInput';
-export * from './MaxAuthorsSlider';
+export * from './MaxAuthorsField';
