+++ src/components/Settings/Export/index.ts
@@ -1,7 +1,7 @@
 export * from './Description';
 export * from './ExportFormatSelect';
 export * from './KeyFormatInputApply';
-export * from './MaxAuthorSlider';
+export * from './MaxAuthorField';
 export * from './NumberSlider';
 export * from './SampleTextArea';
 export * from './Tabs';
